Use more .lua subfiles for different plugin modules
