+++ scripts/vscripts/util/functions.lua
@@ -1,5 +1,10 @@
 -- put your fancy global functions here for you (and others) to use, i guess
 
+-- Apparently entity indices take up the first 14 bits of an EHandle, need more testing to really verify this
+function EHandleToHScript(iPawnId)
+    return EntIndexToHScript(bit.band(iPawnId, 0x3FFF))
+end
+
 -- shuffles positions of elements in an array
 -- usable only for array type of tables (when keys are not strings)
 function table.shuffle(tbl)
